Sensitivity alpha-curve associated to MLP function of an input variable

Obtain sensitivity alpha-curve associated to MLP function obtained from the sensitivities returned by SensAnalysisMLP of an input variable.

Usage

AlphaSensCurve(sens, tol = NULL, max_alpha = 100)

Arguments

sens

raw sensitivities of the MLP output with respect to input variable.

tol

difference between M_alpha and maximum sensitivity of the sensitivity of each input variable

max_alpha

maximum alpha value to analyze

Value

alpha-curve of the MLP function

Examples

# \donttest{
mod <- RSNNS::mlp(simdata[, c("X1", "X2", "X3")], simdata[, "Y"],
                 maxit = 1000, size = 15, linOut = TRUE)

sens <- SensAnalysisMLP(mod, trData = simdata,
                        output_name = "Y", plot = FALSE)

AlphaSensCurve(sens$raw_sens[[1]][,1])
#>   [1] 1.600859 2.024488 2.377516 2.683083 2.950296 3.184139 3.388838 3.568474
#>   [9] 3.726854 3.867346 3.992821 4.105667 4.207850 4.300979 4.386367 4.465087
#>  [17] 4.538020 4.605888 4.669291 4.728724 4.784603 4.837276 4.887040 4.934147
#>  [25] 4.978815 5.021234 5.061570 5.099970 5.136566 5.171475 5.204804 5.236652
#>  [33] 5.267107 5.296251 5.324160 5.350905 5.376551 5.401159 5.424785 5.447483
#>  [41] 5.469303 5.490290 5.510488 5.529939 5.548680 5.566747 5.584174 5.600993
#>  [49] 5.617234 5.632924 5.648091 5.662758 5.676950 5.690688 5.703993 5.716884
#>  [57] 5.729381 5.741501 5.753259 5.764673 5.775755 5.786521 5.796984 5.807156
#>  [65] 5.817048 5.826673 5.836040 5.845160 5.854042 5.862696 5.871130 5.879353
#>  [73] 5.887371 5.895194 5.902827 5.910277 5.917552 5.924657 5.931597 5.938380
#>  [81] 5.945009 5.951490 5.957828 5.964028 5.970094 5.976031 5.981842 5.987531
#>  [89] 5.993103 5.998561 6.003907 6.009147 6.014282 6.019317 6.024253 6.029094
#>  [97] 6.033842 6.038501 6.043072 6.047558
# }